% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the location of fatal motor vehicle accidents in selected state in the US by given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numerical number representing the state you would like to plot}

\item{year}{The year of which you would like to plot accidents from}
}
\value{
returns a map of the requested state with dots representing each accident
}
\description{
Map the location of fatal motor vehicle accidents in selected state in the US by given year
}
\note{
Will return an error if invalid state number given

Will return an error if a year with no data is given

Will return a message saying "no accidents to plot" if there were no
accidents in the state in the year requested
}
\examples{
\donttest{
fars_map_state(15, 2014)
fars_map_state(35, 2014)
}
}
