% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read CSV by Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer or a vector of integers for year(s) to take data from}
}
\value{
Returns a tibble with two columns MONTH and Year.
}
\description{
reads in years from a data frame.
}
\note{
Will return an error if given year is invalid.
}
