% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of people killed in motor vehicle accidents in the US by given year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector of years you want to summarize data from.}
}
\value{
This function returns a number of envents summary table as tibble grouped by month and year
}
\description{
Summarize the number of people killed in motor vehicle accidents in the US by given year and month
}
\examples{
\donttest{
fars_summarize_years(2014)
fars_summarize_years(c(2013,2014,2015))
}
}
