% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create character string containing a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year for which the file name is created}
}
\value{
a CSV filename as a character string
}
\description{
Create character string containing a file name
}
\examples{
make_filename(2014)

}
